Hoist form components up a directory
