+++ components/OverflowMenu/stories.react.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import classNames from 'classnames/bind'
-import Button from 'components/forms/button/Button/react'
+import Button from 'components/Button/react'
 import OverflowMenu from './react'
 import storyStyles from './stories-style.module.scss'
