+++ components/WizardModal/react.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import onClickOutside from 'react-onclickoutside'
 import classNames from 'classnames/bind'
-import { React as TextField } from 'components/forms/TextField'
+import { React as TextField } from 'components/TextField'
 import ModalHeader from './react/ModalHeader'
 import ModalBody from './react/ModalBody'
 import ModalFooter from './react/ModalFooter'
