--- components/forms/button/More/style.scss
@@ -1,3 +1,0 @@
-.oval {
-  box-shadow: inset 0 0.5px 0 0 rgba(0, 0, 0, 0.08);
-}
