+++ packages/external-ui-react/src/index.js
@@ -1,7 +1,7 @@
-export { React as Button } from 'components/forms/button/Button'
-export { React as Checkbox } from 'components/forms/Checkbox'
+export { React as Button } from 'components/Button'
+export { React as Checkbox } from 'components/Checkbox'
 export { React as FilterableSearch } from 'components/forms/FilterableSearch'
-export { React as Input, React as TextField } from 'components/forms/TextField'
+export { React as Input, React as TextField } from 'components/TextField'
 export { React as PaginationDots } from 'components/PaginationDots'
 export { React as ModalBackdrop } from 'components/ModalBackdrop'
 export { React as OverflowMenu } from 'components/menu/OverflowMenu'
